+++ apps/designer/app/designer/features/sidebar-left/panels/asset-manager/index.ts
@@ -1,0 +1,2 @@
+export * from "./add-an-asset-form";
+export * from "./thumbnail";
