+++ apps/designer/app/designer/features/sidebar-left/panels/asset-manager/thumbnail.tsx
