+++ apps/designer/app/designer/features/sidebar-left/panels/asset-manager/uploading-animation.tsx
